refactor: sweep the message services against the naming conventions
`callParticipantMap.ts` exported `callSessionParticipantMap` — the filename
dropped the word that says what the map is keyed by, which is the mismatch the
filename-is-the-export rule exists to stop, and the one the naming skill had been
citing from this very file. The file takes the export's name; the skill's example
now reads as the shape rather than as a place to go look.

`joinLiveKitCall` spelled the product `livekit` in the one place it was a local
rather than an identifier the repo authored, and named the join's own answer
`result` beside it.

Everything else in the unit reads correctly, including the two `create*`
participant helpers: one builds the `CallParticipant` and the other registers it
in the session map, so they are two operations rather than two spellings.
