+++ .agents/ledgers/naming.md
@@ -18,7 +18,7 @@
 | `server/trpc/routers` — `achievement`, `app`, `clicker`, `dungeons`, `notification`, `pushSubscription`, `session`, `storage`               | 2026-09-03 | a caller-scoped read is `readMy*`, never `readOwn*`                                             |
 | `server/trpc/{guards,procedure,plugins,middleware}`, `context.ts`                                                                           | 2026-09-03 | a guard that throws is `assert*`; `is*` is a stored boolean and nothing else                    |
 | `server/composables`, `server/api`, `server/routes`                                                                                         | 2026-09-03 | `get*` vs `read*` on the server side                                                            |
-| `server/services/message`                                                                                                                   | —          |                                                                                                 |
+| `server/services/message`                                                                                                                   | 2026-09-03 | a product's own casing carries into a local — `liveKit`, never `livekit`                        |
 | `server/services/resource`                                                                                                                  | 2026-09-03 |                                                                                                 |
 | `server/services` — `room`, `role`, `user`, `friend`, `post`                                                                                | 2026-09-03 |                                                                                                 |
 | `server/services` — `blueprint`, `program`, `survey`, `dataset`, `dashboard`, `emailEditor`                                                 | 2026-09-03 | a constant map or set stays PascalCase; a scalar is `SCREAMING_SNAKE`                           |
