+++ packages/app/server/trpc/routers/call/knocker.ts
@@ -4,7 +4,7 @@ import type { Context } from "@@/server/trpc/context";
 import { on } from "@@/server/services/events/on";
 import { callAdmittedParticipantMap } from "@@/server/services/message/call/callAdmittedParticipantMap";
 import { callKnockerMap } from "@@/server/services/message/call/callKnockerMap";
-import { callSessionParticipantMap } from "@@/server/services/message/call/callParticipantMap";
+import { callSessionParticipantMap } from "@@/server/services/message/call/callSessionParticipantMap";
 import { createParticipant } from "@@/server/services/message/call/createParticipant";
 import { requireCallSession } from "@@/server/services/message/call/requireCallSession";
 import { requireKnockerCallSession } from "@@/server/services/message/call/requireKnockerCallSession";
